Move field classes to separate module.
